requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
astor=0.8.1=pyh9f0ad1d_0
attrs=21.2.0=pyhd8ed1ab_0
blas=1.0=mkl
bokeh=2.3.3=py39h6e9494a_0
bottleneck=1.3.2=py39hae3b40e_3
ca-certificates=2021.5.30=h033912b_0
certifi=2021.5.30=py39h6e9494a_0
cftime=1.5.0=py39hc89836e_0
click=8.0.1=py39h6e9494a_0
cloudpickle=1.6.0=py_0
codespell=2.1.0=pyhd8ed1ab_0
curl=7.71.1=hb0a8c7a_1
cytoolz=0.11.0=py39h89e85a6_3
dask=2021.7.2=pyhd8ed1ab_0
dask-core=2021.7.2=pyhd8ed1ab_0
distributed=2021.7.2=py39h6e9494a_0
doc8=0.9.0=pyhd8ed1ab_0
docutils=0.17.1=py39h6e9494a_0
flake8=3.9.2=pyhd8ed1ab_0
flake8-bugbear=21.4.3=pyhd8ed1ab_0
flake8-builtins=1.5.3=pyh9f0ad1d_0
flake8-comprehensions=3.6.1=pyhd8ed1ab_0
flake8-continuation=1.0.5=pyhd8ed1ab_0
flake8-copyright=0.2.2=py_0
flake8-docstrings=1.6.0=pyhd8ed1ab_0
flake8-isort=4.0.0=pyhd8ed1ab_1
flake8-mutable=1.2.0=py_1
flake8-polyfill=1.0.2=py_0
flake8-print=4.0.0=pyhd8ed1ab_0
flake8-quotes=3.3.0=pyhd8ed1ab_0
flake8-rst-docstrings=0.2.3=pyhd8ed1ab_0
flake8-simplify=0.14.1=pyhd8ed1ab_0
freetype=2.10.4=h4cff582_1
fsspec=2021.7.0=pyhd8ed1ab_0
grpcio=1.39.0=pypi_0
grpcio-tools=1.39.0=pypi_0
hdf4=4.2.13=h39711bb_2
hdf5=1.10.6=hdbbcd12_0
heapdict=1.0.1=py_0
importlib-metadata=4.8.1=py39h6e9494a_0
importlib_metadata=4.8.1=hd8ed1ab_0
intel-openmp=2021.3.0=hecd8cb5_3375
isort=5.9.3=pyhd8ed1ab_0
jinja2=3.0.1=pyhd8ed1ab_0
jpeg=9d=hbcb3906_0
krb5=1.18.2=h75d18d8_0
lcms2=2.12=h577c468_0
libcurl=7.71.1=h8a08a2b_1
libcxx=10.0.0=1
libedit=3.1.20191231=h0678c8f_2
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libnetcdf=4.6.1=hfd9a460_4
libpng=1.6.37=h7cec526_2
libssh2=1.9.0=h52ee1ee_6
libtiff=4.2.0=h87d7836_0
libwebp-base=1.2.0=h0d85af4_2
llvm-openmp=12.0.1=hda6cdc1_1
locket=0.2.0=py_2
lz4-c=1.9.2=h4a8c4bd_1
markupsafe=2.0.1=py39h89e85a6_0
mccabe=0.6.1=pypi_0
mkl=2021.3.0=hecd8cb5_517
mkl-service=2.4.0=py39h9ed2024_0
mkl_fft=1.3.0=py39h4a7008c_2
mkl_random=1.2.2=py39hb2f4e1b_0
msgpack-python=1.0.2=py39hf7b0b51_1
ncurses=6.2=h0a44026_1
netcdf4=1.5.7=py39h93ad9c5_0
numpy=1.20.3=py39h4b4dc7a_0
numpy-base=1.20.3=py39he0bd621_0
olefile=0.46=pyh9f0ad1d_1
openssl=1.1.1k=h0d85af4_1
packaging=21.0=pyhd8ed1ab_0
pandas=1.2.4=py39h23ab428_0
partd=1.2.0=pyhd8ed1ab_0
pbr=5.6.0=pyhd8ed1ab_0
pep8-naming=0.12.1=pyhd8ed1ab_0
pillow=7.2.0=py39h31b929e_2
pip=21.2.2=py39hecd8cb5_0
protobuf=3.17.3=pypi_0
psutil=5.8.0=py39h89e85a6_1
pycodestyle=2.7.0=pyhd8ed1ab_0
pydocstyle=6.1.1=pyhd8ed1ab_0
pyflakes=2.3.1=pyhd8ed1ab_0
pygments=2.10.0=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
python=3.9.6=h88f2d9e_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytz=2021.1=pyhd8ed1ab_0
pyyaml=5.4.1=py39hcbf5805_0
readline=8.1=h9ed2024_0
restructuredtext_lint=1.3.2=pyhd3deb0d_0
scipy=1.6.2=py39hd5f7400_1
setuptools=52.0.0=py39hecd8cb5_0
six=1.16.0=pyh6c4a22f_0
snowballstemmer=2.1.0=pyhd8ed1ab_0
sortedcontainers=2.4.0=pyhd8ed1ab_0
sqlite=3.36.0=hce871da_0
stevedore=3.4.0=py39h6e9494a_0
tblib=1.7.0=pyhd8ed1ab_0
testfixtures=6.18.1=pyhd8ed1ab_0
tk=8.6.10=hb0a8c7a_0
toolz=0.11.1=py_0
tornado=6.1=py39h89e85a6_1
typing_extensions=3.10.0.0=pyha770c72_0
tzdata=2021a=h52ac0ba_0
wheel=0.36.2=pyhd3eb1b0_0
xarray=0.19.0=pyhd8ed1ab_1
xz=5.2.5=h1de35cc_0
yaml=0.2.5=haf1e3a3_0
zict=2.0.0=py_0
zipp=3.5.0=pyhd8ed1ab_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.5=h41d2c2f_0
